Clear accumulated wattseconds prior to EVSE_STATE_DISABLED
https://openev.freshdesk.com/support/discussions/topics/6000046121?page=1
has a report from a user that disabling and re-enabling the unit
will cause the lifetime watt-hour display to increase by the
last session's watt-second total on each disable/enable cycle.

This is because going into the disabled state does not clear the
watt-second counter, and when we re-enable it and transition to
EVSE_STATE_A, it will add g_WattSeconds to g_WattHours_accumulated
each time.

Fix this by setting wattseconds to 0 prior to disabling the unit,
similar to what is done prior to Sleep().

Signed-off-by: Eric Sandeen <sandeen@sandeen.net>