Update open_evse.ino
If Real-Time Clock isn't set, and if we're not setting via the wifi module then the YEAR_MIN should be a bit more recent to save button pushes.
Perhaps YEAR_MAX for the pushbutton menu interface should be more than 10 years later? What's the expected lifespan of this system before we'd expect firmware updates to be applied?